Remove _edit_attributes.html.haml
The user will need to go to the raw attributes page to change the list
of devices. So instead of having a completely useless custom attributes
page (with just a link to the raw page), force the display of the raw
page by having no custom page.

Workaround for https://bugzilla.novell.com/show_bug.cgi?id=778088